ENH: Dynamically add relative time difference first column during editing
The absolute times make it very hard to tweak durations or add a comment in the middle.
https://github.com/asciinema/asciinema/issues/594 explains that the v1 format actually had deltas.
In Vim, it's easy to insert a "virtual" column on reading the buffer and removing it again on writing. The time offset take precedence over the original absolute times, so by modifying them the timing can be easily tweaked.

System-info: Ubuntu 20.04.6 LTS, x86_64
Platform-info: Vim 9.1.1029